context.rsの作成/切り出し
- あと、with_helperのコメントの修正
